# api/openapi.yaml
openapi: 3.0.0
info:
  description: The API for the simple containerisation solution Podinate.
  title: Podinate API
  version: 0.0.1
servers:
- description: Production server
  url: https://api.podinate.com/v0
- description: Development Server
  url: http://localhost:3001/v0
security:
  - APIKeyAuth: []
...
paths:
  /account:
    get:
      summary: Get information about the current account.
      description: Accounts are logically separate, distinct containers for different customers or teams. For example you might have an account for your marketing and product teams, or you might have an account for each client to separate their workloads. Accounts can have their own security settings applied.
      tags:
      - Account
      parameters:
        - $ref: '#/components/parameters/account'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
          description: Information about the given account
        "404":
          $ref: "#/components/responses/404Error"
        "500":
          $ref: "#/components/responses/500Error"
        "409":
          $ref: "#/components/responses/409Error"
    post:
      summary: Create a new account
      description: Create a new logically separate account, for example for a given client or team.
      tags:
      - Account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
      responses:
        "201":
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        "400":
          $ref: "#/components/responses/400Error"
        "500":
          $ref: "#/components/responses/500Error"
    put:
      summary: Update an existing account
      description: Update an existing account, for example to change the display name
      tags:
      - Account
      parameters:
      - $ref: '#/components/parameters/account'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
      responses:
        "200":
          description: Account updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        "400":
          $ref: "#/components/responses/400Error"
        "500":
          $ref: "#/components/responses/500Error"
          description: The server was not able to update your account
    delete:
      summary: Delete the account and all associated resources!
      description: Deletes the account and all associated resources! This is irreversible and may be used for example when a client project has finished or a migration is complete. 
      tags:
      - Account
      parameters:
      - $ref: '#/components/parameters/account'
      responses:
        "202":
          description: Podinate receieved your request to delete the account - it will be processed in the background. 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        "400":
          $ref: "#/components/responses/400Error"
        "404":
          $ref: "#/components/responses/404Error"
        "500":
          $ref: "#/components/responses/500Error"
...
  /project/{project_id}/pod/{pod_id}/logs:
    get:
      summary: Get the logs for a pod
      description: Get the logs for a pod
      tags:
      - Pod
      x-speakeasy-entity-operation: Pod#logs
      parameters:
      - explode: false
        in: query
        name: lines
        description: The number of lines to return
        schema:
          type: integer
          default: 20
      - explode: false
        in: query
        name: follow
        description: Whether to keep the connection open and continue streaming the logs
        schema:
          type: boolean
          default: false
      - explode: false
        in: path
        name: project_id
        required: true
        schema:
          description: The project to look at
          example: hello-world
          type: string
        style: simple
      - explode: false
        in: path
        name: pod_id
        required: true
        schema:
          description: The pod to look at
          example: hello-world
          type: string
        style: simple
        x-speakeasy-match: id
      - $ref: '#/components/parameters/account'
      responses:
        "200":
          content:
            text/plain:
              schema:
                type: string
                example: "127.0.0.1 - peter [9/Feb/2017:10:34:12 -0700] \"GET /sample-image.png HTTP/2\" 200 1479"
          description: The logs for the given pod
        "400":
          $ref: "#/components/responses/400Error"
        "404":
          $ref: "#/components/responses/404Error"
        "500":
          $ref: "#/components/responses/500Error"
  /project/{project_id}/pod/{pod_id}/exec:
    post:
      summary: Execute a command in a pod
      description: Execute a command in a pod
      tags:
      - Pod
      x-speakeasy-entity-operation: Pod#exec
      parameters:
      - explode: false
        in: path
        name: project_id
        required: true
        schema:
          description: The project to look at
          example: hello-world
          type: string
        style: simple
      - explode: false
        in: path
        name: pod_id
        required: true
        schema:
          description: The pod to look at
          example: hello-world
          type: string
        style: simple
        x-speakeasy-match: id
      - $ref: '#/components/parameters/account'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                command:
                  type: array
                  items:
                    type: string
                  description: The command to execute
                  example: ["ls", "/"]
              required:
                - command
      responses:
        "200":
          content:
            text/plain:
              schema:
                type: string
                example: "bin dev etc home lib lib64 media mnt opt proc root run sbin srv sys tmp usr var"
          description: The output of the command
        "400":
          $ref: "#/components/responses/400Error"
        "404":
          $ref: "#/components/responses/404Error"
        "500":
          $ref: "#/components/responses/500Error"
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The command could not be executed due to a permissions error
...
components:
  securitySchemes:
    APIKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  parameters:
    account:
      description: The account to use for the request
      example: my-account
      explode: false
      in: header
      name: account
      required: true
      schema:
        type: string
      style: simple
...
  responses:
    "400Error":
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: "Request issued incorrectly, for example missing parameters or\
        \ wrong endpoint"
    "404Error":
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: "The requested resource was not found"
    "500Error":
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: An internal server error
    "409Error":
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: "The request could not be completed due to a conflict with the\
        \ current state of the resource"
  schemas:
    Account:
      example:
        id: "blog-prod"
        name: "Company Blog - Prod"
      properties:
        id:
          description: The short name (slug/url) of the account. The account ID is globally unique and cannot be changed.
          example: "blog-prod"
          type: string 
        name:
          description: The human readable name of the account, used for display purposes.
          example: "Tim's Widgets Company Blog - Prod"
          type: string
        resource_id:
          description: The global Resource ID of the account
          example: account:podinate-blog-prod
          type: string
          x-speakeasy-param-readonly: true
...
    "Error":
      properties:
        code:
          default: 400
          description: The http code of the error message (400)
          example: 400
          type: number
        message:
          description: The human-readable description of what went wrong
          example: Missing required parameter 'account'
          type: string
      type: object